requirements.txt
matplotlib==3.3.4
nest-asyncio==1.5.1
numpy==1.18.5
tensorboard==2.4.1
tensorflow==2.3.2
tensorflow-federated==0.17.0
tensorflow-addons==0.11.2
SimpleITK==2.0.2
attrs==19.3.0
